+++ src/main/java/com/techathome/repository/CartDetailRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CartDetailRepository extends JpaRepository <CartDetail, Integer> {
+public interface CartDetailRepository extends JpaRepository <CartDetail, Long> {
 
 	Optional<CartDetail> findByCartCartIdAndProductProductId(Long cartId, Long productId);
 }
